... servo_util.cpp

+++ main/include/servo_ctrl/servo_util.cpp
@@ -153,7 +153,7 @@ void Servo_comm::task()
 
 		target_rc = target.load(); if(nop){}
 		target_rc = REVERSED ? (float)ROT_MAX_DEG-target_rc:target_rc;
-		if (fabs(target_rc - rot_current) < 5e-1 || target_rc > (uint32_t)ROT_MAX_DEG)
+		if (fabs(target_rc - rot_current) < 5e-1 || target_rc > (float)ROT_MAX_DEG)
 			continue;
 
 		rot_relat = (float)(target_rc - rot_current) / ROT_MAX_DEG;
